+++ server/src/handlers/common.ts
@@ -1,0 +1,1 @@
+export const streamIdCookieName = 'JustCastId'
