+++ server/src/handlers/index.ts
@@ -1,0 +1,3 @@
+export { api } from './api'
+export { websocket } from './websocket'
+export { files } from './files'
